Convert test utils to classes
