+++ src/renderer/App/AppBar/SearchBox/tests/SearchBox.test.ts
@@ -1,10 +1,10 @@
 import { fireEvent } from "@testing-library/react";
-import { findByTestId } from "_tu/findByTestId";
+import { TestElement } from "_tu/TestElement";
 import * as blockFixtures from "_tu/fixtures/blockFixtures";
-import { initializeElectronCode } from "_tu/initializeElectronCode";
-import { renderAppWithStore } from "_tu/renderAppWithStore";
-import { waitWithTime } from "_tu/smallUtils";
-import { startMockRpcServer } from "_tu/startMockRpcServer";
+import { MockElectron } from "_tu/MockElectron";
+import { appWithStore } from "_tu/AppWithStore";
+import { Utils } from "_tu/Utils";
+import { MockRpcServer } from "_tu/MockRpcServer";
 import { testIds } from "_r/testIds";
 
 jest.mock("_f/featureFlags", () => ({
@@ -17,34 +17,38 @@ jest.mock("_f/featureFlags", () => ({
 describe("SearchBox", () => {
   describe("Search flow", () => {
     beforeAll(async () => {
-      startMockRpcServer();
-      initializeElectronCode();
-      await renderAppWithStore();
+      MockRpcServer.start();
+      MockElectron.start();
+      await appWithStore.render();
     });
 
     test("app loads with the search box visible", async () => {
       /**
        * The app loads with the search box visible
        */
-      expect(await findByTestId("searchInputField")).toBeVisible();
+      expect(await TestElement.findByTestId("searchInputField")).toBeVisible();
     });
 
     test("search for a block by height", async () => {
       /**
        * We will start by searching for a block by height
        */
-      fireEvent.change(await findByTestId("searchInputField"), {
+      fireEvent.change(await TestElement.findByTestId("searchInputField"), {
         target: { value: blockFixtures.blockFixture19.height },
       });
 
-      fireEvent.keyUp(await findByTestId("searchInputField"), { keyCode: 13 });
+      fireEvent.keyUp(await TestElement.findByTestId("searchInputField"), {
+        keyCode: 13,
+      });
 
       /**
        * `h1` is showing the block height of `blockFixture2` because we searched
        * for it
        */
-      await waitWithTime(async () =>
-        expect(await findByTestId(testIds.blockDetailsH1)).toHaveTextContent(
+      await Utils.waitWithTime(async () =>
+        expect(
+          await TestElement.findByTestId(testIds.blockDetailsH1),
+        ).toHaveTextContent(
           `#${blockFixtures.blockFixture19.height.toLocaleString()}`,
         ),
       );
@@ -54,35 +58,43 @@ describe("SearchBox", () => {
       /**
        * We can now try searching for blockFixture3 by hash
        */
-      fireEvent.change(await findByTestId("searchInputField"), {
+      fireEvent.change(await TestElement.findByTestId("searchInputField"), {
         target: { value: blockFixtures.blockFixture18.hash },
       });
 
-      fireEvent.keyUp(await findByTestId("searchInputField"), { keyCode: 13 });
+      fireEvent.keyUp(await TestElement.findByTestId("searchInputField"), {
+        keyCode: 13,
+      });
 
-      await waitWithTime(async () =>
-        expect(await findByTestId(testIds.blockDetailsH1)).toHaveTextContent(
+      await Utils.waitWithTime(async () =>
+        expect(
+          await TestElement.findByTestId(testIds.blockDetailsH1),
+        ).toHaveTextContent(
           `#${blockFixtures.blockFixture18.height.toLocaleString()}`,
         ),
       );
     });
 
     test("searching by transaction", async () => {
-      fireEvent.change(await findByTestId("searchInputField"), {
+      fireEvent.change(await TestElement.findByTestId("searchInputField"), {
         target: { value: blockFixtures.blockFixture18.tx[2] },
       });
 
-      fireEvent.keyUp(await findByTestId("searchInputField"), { keyCode: 13 });
+      fireEvent.keyUp(await TestElement.findByTestId("searchInputField"), {
+        keyCode: 13,
+      });
 
-      expect(await findByTestId("transactionDetails")).toBeVisible();
+      expect(
+        await TestElement.findByTestId("transactionDetails"),
+      ).toBeVisible();
     });
 
     test("it does not do anything if we modify the search field but try to submit with a key other than enter, like shift", async () => {
-      fireEvent.change(await findByTestId("searchInputField"), {
+      fireEvent.change(await TestElement.findByTestId("searchInputField"), {
         target: { value: blockFixtures.blockFixture19.hash },
       });
 
-      fireEvent.keyUp(await findByTestId("searchInputField"), {
+      fireEvent.keyUp(await TestElement.findByTestId("searchInputField"), {
         keyCode: 16 /* shift */,
       });
 
@@ -91,25 +103,31 @@ describe("SearchBox", () => {
        * the previous test is still showing. Pressing shift didn't
        * trigger the search.
        */
-      await waitWithTime(async () =>
-        expect(await findByTestId(testIds.blockDetailsH1)).toHaveTextContent(
+      await Utils.waitWithTime(async () =>
+        expect(
+          await TestElement.findByTestId(testIds.blockDetailsH1),
+        ).toHaveTextContent(
           `#${blockFixtures.blockFixture18.height.toLocaleString()}`,
         ),
       );
     });
 
     test("it does not do anything when the search string does not return a block", async () => {
-      fireEvent.change(await findByTestId("searchInputField"), {
+      fireEvent.change(await TestElement.findByTestId("searchInputField"), {
         target: { value: "🕺" },
       });
 
-      fireEvent.keyUp(await findByTestId("searchInputField"), { keyCode: 13 });
+      fireEvent.keyUp(await TestElement.findByTestId("searchInputField"), {
+        keyCode: 13,
+      });
 
       /**
        * Same block is still displayed. Search didn't cause a change.
        */
-      await waitWithTime(async () =>
-        expect(await findByTestId(testIds.blockDetailsH1)).toHaveTextContent(
+      await Utils.waitWithTime(async () =>
+        expect(
+          await TestElement.findByTestId(testIds.blockDetailsH1),
+        ).toHaveTextContent(
           `#${blockFixtures.blockFixture18.height.toLocaleString()}`,
         ),
       );
