+++ src/renderer/App/Explorer/BlockList/BlockList.test.ts
@@ -1,9 +1,9 @@
-import { findByTestId } from "_tu/findByTestId";
-import { startMockRpcServer } from "_tu/startMockRpcServer";
-import { initializeElectronCode } from "_tu/initializeElectronCode";
-import { renderAppWithStore } from "_tu/renderAppWithStore";
+import { TestElement } from "_tu/TestElement";
+import { MockRpcServer } from "_tu/MockRpcServer";
+import { MockElectron } from "_tu/MockElectron";
+import { appWithStore } from "_tu/AppWithStore";
 import * as blockFixtures from "_tu/fixtures/blockFixtures";
-import { userEvent, waitWithTime } from "_tu/smallUtils";
+import { Utils } from "_tu/Utils";
 import { fireEvent, wait } from "@testing-library/react";
 import { testIds } from "_r/testIds";
 
@@ -16,47 +16,53 @@ jest.mock("_f/featureFlags", () => ({
 
 describe("BlockList", () => {
   beforeAll(async () => {
-    startMockRpcServer();
-    initializeElectronCode();
-    await renderAppWithStore();
+    MockRpcServer.start();
+    MockElectron.start();
+    await appWithStore.render();
   });
 
   it("indicates to the user when they are at the top of the block chain", async () => {
-    expect(await findByTestId("depthTopLink")).toHaveTextContent(
+    expect(await TestElement.findByTestId("depthTopLink")).toHaveTextContent(
       "Top of the block chain",
     );
 
-    expect(await findByTestId("depthBottomLink")).toHaveTextContent(
+    expect(await TestElement.findByTestId("depthBottomLink")).toHaveTextContent(
       `There is ${blockFixtures.blockFixture1.height.toLocaleString()} lower block`,
     );
   });
 
   it("indicates to the user how many blocks there are higher than the highest block in the block list", async () => {
-    fireEvent.change(await findByTestId("searchInputField"), {
+    fireEvent.change(await TestElement.findByTestId("searchInputField"), {
       target: { value: blockFixtures.blockFixture0.height },
     });
 
-    fireEvent.keyUp(await findByTestId("searchInputField"), { keyCode: 13 });
+    fireEvent.keyUp(await TestElement.findByTestId("searchInputField"), {
+      keyCode: 13,
+    });
 
-    await waitWithTime(async () =>
-      expect(await findByTestId(testIds.blockDetailsH1)).toHaveTextContent(
+    await Utils.waitWithTime(async () =>
+      expect(
+        await TestElement.findByTestId(testIds.blockDetailsH1),
+      ).toHaveTextContent(
         `#${blockFixtures.blockFixture0.height.toLocaleString()}`,
       ),
     );
 
-    expect(await findByTestId("depthTopLink")).toHaveTextContent(
+    expect(await TestElement.findByTestId("depthTopLink")).toHaveTextContent(
       "There is 1 higher block",
     );
-    expect(await findByTestId("depthBottomLink")).toHaveTextContent(
+    expect(await TestElement.findByTestId("depthBottomLink")).toHaveTextContent(
       "Genesis of the block chain", // praise be to satoshi!
     );
   });
 
   it("focuses the search box when the depth bottom link search button is pressed", async () => {
-    userEvent.click(await findByTestId("depthBottomLinkSearchButton"));
+    Utils.userEvent.click(
+      await TestElement.findByTestId("depthBottomLinkSearchButton"),
+    );
 
     await wait(async () =>
-      expect(await findByTestId("searchInputField")).toHaveFocus(),
+      expect(await TestElement.findByTestId("searchInputField")).toHaveFocus(),
     );
   });
 });
