+++ src/renderer/App/RpcIssueDialog/tests/RpcIssueDialog.missingBitcoinConf.test.ts
@@ -1,7 +1,7 @@
-import { findByTestId } from "_tu/findByTestId";
-import { initializeElectronCode } from "_tu/initializeElectronCode";
-import { renderAppWithStore } from "_tu/renderAppWithStore";
-import { startMockRpcServer } from "_tu/startMockRpcServer";
+import { TestElement } from "_tu/TestElement";
+import { MockElectron } from "_tu/MockElectron";
+import { appWithStore } from "_tu/AppWithStore";
+import { MockRpcServer } from "_tu/MockRpcServer";
 
 jest.mock("_f/featureFlags", () => ({
   __esModule: true,
@@ -12,12 +12,14 @@ jest.mock("_f/featureFlags", () => ({
 
 describe("RpcIssueDialog - missing `bitcoin.conf`", () => {
   beforeAll(async () => {
-    startMockRpcServer();
-    initializeElectronCode({ skipInitializingFilesystem: true }); // Don't add `bitcoin.conf` to the filesystem.
-    await renderAppWithStore();
+    MockRpcServer.start();
+    MockElectron.start({ skipInitializingFilesystem: true }); // Don't add `bitcoin.conf` to the filesystem.
+    await appWithStore.render();
   });
 
   it("brings up the RPC issue dialog because there is no `bitcoin.conf`", async () => {
-    expect(await findByTestId("fixBcoreConnectionDialog")).toBeInTheDocument();
+    expect(
+      await TestElement.findByTestId("fixBcoreConnectionDialog"),
+    ).toBeInTheDocument();
   });
 });
