--- src/testUtils/findByTestId.ts
@@ -1,11 +1,0 @@
-import { screen } from "@testing-library/dom";
-import { testIds } from "../renderer/testIds";
-
-type TestId = keyof typeof testIds;
-
-export const findByTestId = (testId: TestId) => screen.findByTestId(testId);
-
-export const queryByTestId = (testId: TestId) => screen.queryByTestId(testId);
-
-export const findAllByTestId = (testId: TestId) =>
-  screen.findAllByTestId(testId);
